extract argument parser module
