+++ protostar/cli/__init__.py
@@ -1,9 +1,1 @@
 from protostar.cli.activity_indicator import ActivityIndicator
-from protostar.cli.argument_parser_facade import (
-    ArgumentParserFacade,
-    MissingRequiredArgumentException,
-)
-from protostar.cli.cli_app import CLIApp
-from protostar.cli.command import Command
-
-from .config_file_argument_resolver import ConfigFileArgumentResolverProtocol
