+++ protostar/commands/declare_command.py
@@ -5,7 +5,7 @@ from typing import List, Optional
 from starknet_py.net.gateway_client import GatewayClient
 from starknet_py.net.signer import BaseSigner
 
-from protostar.cli.command import Command
+from protostar.argument_parser import Command
 from protostar.cli.network_command_util import NetworkCommandUtil
 from protostar.cli.signable_command_util import SignableCommandUtil
 from protostar.commands.deploy_command import DeployCommand
