+++ protostar/commands/deploy_command.py
@@ -1,7 +1,7 @@
 from logging import Logger
 from typing import List, Optional
 
-from protostar.cli.command import Command
+from protostar.argument_parser import Command
 from protostar.cli.network_command_util import NetworkCommandUtil
 from protostar.starknet_gateway import (
     GatewayFacadeFactory,
