+++ protostar/commands/install/install_command.py
@@ -2,14 +2,14 @@ from logging import Logger
 from pathlib import Path
 from typing import Callable, List, Optional
 
-from protostar.cli import Command
+from protostar.argument_parser import Command
 from protostar.commands.install.install_package_from_repo import (
     install_package_from_repo,
 )
 from protostar.commands.install.pull_package_submodules import pull_package_submodules
+from protostar.io.log_color_provider import LogColorProvider
 from protostar.package_manager import extract_info_from_repo_id
 from protostar.protostar_toml.protostar_project_section import ProtostarProjectSection
-from protostar.io.log_color_provider import LogColorProvider
 
 EXTERNAL_DEPENDENCY_REFERENCE_DESCRIPTION = """- `GITHUB_ACCOUNT_NAME/REPO_NAME[@TAG]`
     - `OpenZeppelin/cairo-contracts@v0.4.0`
