+++ protostar/commands/migrate_command.py
@@ -5,16 +5,16 @@ from typing import List, Optional
 from starknet_py.net.gateway_client import GatewayClient
 from starknet_py.net.signer import BaseSigner
 
-from protostar.commands.build_command import BuildCommand
-from protostar.cli import Command
+from protostar.argument_parser import Command
 from protostar.cli.network_command_util import NetworkCommandUtil
 from protostar.cli.signable_command_util import SignableCommandUtil
+from protostar.commands.build_command import BuildCommand
+from protostar.io.input_requester import InputRequester
+from protostar.io.log_color_provider import LogColorProvider
 from protostar.migrator import Migrator, MigratorExecutionEnvironment
 from protostar.protostar_exception import ProtostarException
 from protostar.starknet import CheatcodeException
 from protostar.starknet_gateway.gateway_facade_factory import GatewayFacadeFactory
-from protostar.io.input_requester import InputRequester
-from protostar.io.log_color_provider import LogColorProvider
 
 
 class MigrateCommand(Command):
