+++ protostar/protostar_cli_test.py
@@ -7,13 +7,12 @@ from typing import Any, List, cast
 import pytest
 from pytest_mock import MockerFixture
 
-from protostar.cli import ArgumentParserFacade, Command
-from protostar.protostar_exception import ProtostarException, ProtostarExceptionSilent
-from protostar.upgrader.latest_version_checker import LatestVersionChecker
+from protostar.argument_parser import ArgumentParserFacade, Command
 from protostar.io.log_color_provider import LogColorProvider
-from protostar.self.protostar_directory import VersionManager
-
 from protostar.protostar_cli import ProtostarCLI
+from protostar.protostar_exception import ProtostarException, ProtostarExceptionSilent
+from protostar.self.protostar_directory import VersionManager
+from protostar.upgrader.latest_version_checker import LatestVersionChecker
 
 
 @pytest.fixture(name="git_version")
